added a module for the displayData function
