+++ diet.html
@@ -79,5 +79,6 @@
         </div>
     </footer>
 </body>
-<script src="./main.js/diet.js"></script>
+<script src="./main.js/modules.js" type="module"></script>
+<script src="./main.js/diet.js" type="module"></script>
 </html>
