+++ index.html
@@ -54,7 +54,7 @@
                         <div class="recipeCardFront" id="recipeResult">
                         <!-- Results (Image and name of recipe) will be appended once called from API -->
                         </div>
-                        <div class="recipeCardBack" id="recipeCardBack">
+                        <div class="recipeCardBack" id="recipeCardBack">.
                             <h2>Ingredients</h2>
                             <div class="ingredientsPlusButton">
                             <ul class="ingredientsList" id="ingredientsList">
@@ -80,5 +80,6 @@
         </div>
     </footer>
 </body>
-<script src="./main.js/script.js"></script>
+<script src="./main.js/modules.js" type="module"></script>
+<script src="./main.js/script.js" type="module"></script>
 </html>
