+++ main.js/script.js
@@ -1,5 +1,7 @@
+import displaydata from "./modules.js";
+
 // Creating Namespace
-recipesApp = {};
+const recipesApp = {};
 
 // Building Array for Cuising Selection
 recipesApp.cuisineOptions = [
@@ -143,7 +145,7 @@ recipesApp.apiCallRecipe = function (recipeIdNumber) {
     })
     .then(function (jsonData) {
             recipesApp.displayIngredients(jsonData);
-            recipesApp.displaydata(jsonData);
+            displaydata(jsonData);
             recipesApp.cardClick.style.display = 'block';
         })
         .catch(function (error) {
@@ -165,36 +167,37 @@ recipesApp.displayIngredients = function(recipesData) {
 }
 
 // Displaying data on page from recipesApp.apiCallRecipe on page
-recipesApp.displaydata = function (recipesData) {
-    const outerDiv = document.querySelector('#recipeResult');
-    const recipeInstructionsDiv = document.querySelector("#recipeInstructions");
-    const newHeading = document.createElement('h2');
-    const imageItem = document.createElement('img');
-    const articleItem = document.createElement("article");
-    // Clearing Div so new recipe can be displayed
-    outerDiv.innerHTML = '';
-    recipeInstructionsDiv.innerHTML = "";
-    // assigning the information from the object to HTML
-    newHeading.textContent = recipesData.title;
-    imageItem.src = recipesData.image;
-    articleItem.textContent = recipesData.instructions;
-      // Appending and formatting everything
-    outerDiv.innerHTML = `
-    <div class="imgContainer"><img src=${imageItem.src} alt="Image of ${newHeading.textContent}"></div>
-    <h2>${newHeading.textContent}</h2>
-    <div class="cookingTimeServings">
-        <div class="cookTime"><i class="fa-regular fa-clock"></i>Cook Time<p>${recipesData.readyInMinutes} mins</p></div>
-        <div class="servings"><i class="fa-regular fa-user"></i>Serves<p>${recipesData.servings}</p></div>
-    </div>
-    <button class="ingredientsButton" id="ingredientsButton">Ingredients</button>
-    `;
+// displaydata(recipesData);
+// recipesApp.displaydata = function (recipesData) {
+//     const outerDiv = document.querySelector('#recipeResult');
+//     const recipeInstructionsDiv = document.querySelector("#recipeInstructions");
+//     const newHeading = document.createElement('h2');
+//     const imageItem = document.createElement('img');
+//     const articleItem = document.createElement("article");
+//     // Clearing Div so new recipe can be displayed
+//     outerDiv.innerHTML = '';
+//     recipeInstructionsDiv.innerHTML = "";
+//     // assigning the information from the object to HTML
+//     newHeading.textContent = recipesData.title;
+//     imageItem.src = recipesData.image;
+//     articleItem.textContent = recipesData.instructions;
+//       // Appending and formatting everything
+//     outerDiv.innerHTML = `
+//     <div class="imgContainer"><img src=${imageItem.src} alt="Image of ${newHeading.textContent}"></div>
+//     <h2>${newHeading.textContent}</h2>
+//     <div class="cookingTimeServings">
+//         <div class="cookTime"><i class="fa-regular fa-clock"></i>Cook Time<p>${recipesData.readyInMinutes} mins</p></div>
+//         <div class="servings"><i class="fa-regular fa-user"></i>Serves<p>${recipesData.servings}</p></div>
+//     </div>
+//     <button class="ingredientsButton" id="ingredientsButton">Ingredients</button>
+//     `;
 
-    recipeInstructionsDiv.innerHTML = 
-    `<article class="recipeInfo">${articleItem.textContent}</article>
-    <p class="recipeLink">  For the original recipe, click the link 
-    <a href="${recipesData.sourceUrl}">here</a> `;
+//     recipeInstructionsDiv.innerHTML = 
+//     `<article class="recipeInfo">${articleItem.textContent}</article>
+//     <p class="recipeLink">  For the original recipe, click the link 
+//     <a href="${recipesData.sourceUrl}">here</a> `;
 
-}
+// }
 
 // Error Handling no response from API
 recipesApp.errorHandle = function () {
