+++ main.js/search.js
@@ -1,5 +1,8 @@
+import displaydata from "./modules.js";
+
+
 // Creating Namespace
-recipesApp = {};
+const recipesApp = {};
 
 // Querying global variables from HTML and saving as variable
 recipesApp.cardClick = document.querySelector('#cardHolder');
@@ -70,7 +73,7 @@ recipesApp.apiCallRecipe = function (recipeIdNumber) {
     })
     .then(function (jsonData) {
             recipesApp.displayIngredients(jsonData);
-            recipesApp.displaydata(jsonData);
+            displaydata(jsonData);
             recipesApp.cardClick.style.display = 'block';
         })
         .catch(function (error) {
@@ -89,37 +92,6 @@ recipesApp.displayIngredients = function(recipesData) {
         ulItem.appendChild(listItem);
 
     });
-}
-
-// Displaying data on page from recipesApp.apiCallRecipe on page
-recipesApp.displaydata = function (recipesData) {
-    const outerDiv = document.querySelector('#recipeResult');
-    const recipeInstructionsDiv = document.querySelector("#recipeInstructions");
-    const newHeading = document.createElement('h2');
-    const imageItem = document.createElement('img');
-    const articleItem = document.createElement("article");
-    // Clearing Div so new recipe can be displayed
-    outerDiv.innerHTML = '';
-    // assigning the information from the object to HTML
-    newHeading.textContent = recipesData.title;
-    imageItem.src = recipesData.image;
-    articleItem.textContent = recipesData.instructions;
-      // Appending and formatting everything
-    outerDiv.innerHTML = `
-    <div class="imgContainer"><img src=${imageItem.src} alt="Image of ${newHeading.textContent}"></div>
-    <h2>${newHeading.textContent}</h2>
-    <div class="cookingTimeServings">
-        <div class="cookTime"><i class="fa-regular fa-clock"></i>Cook Time<p>${recipesData.readyInMinutes} mins</p></div>
-        <div class="servings"><i class="fa-regular fa-user"></i>Serves<p>${recipesData.servings}</p></div>
-    </div>
-    <button class="ingredientsButton" id="ingredientsButton">Ingredients</button>
-    `;
-
-    recipeInstructionsDiv.innerHTML = 
-    `<article class="recipeInfo">${articleItem.textContent}</article>
-    <p class="recipeLink">  For the original recipe, click the link 
-    <a href="${recipesData.sourceUrl}">here</a> `;
-
 }
 
 // Error Handling no response from API
