+++ search.html
@@ -77,5 +77,6 @@
         </div>
     </footer>
 </body>
-<script src="./main.js/search.js"></script>
+<script src="./main.js/modules.js" type="module"></script>
+<script src="./main.js/search.js" type="module"></script>
 </html>
